Edited file 'python-connectors/xzibit_plugins/connector.py'

+++ python-connectors/xzibit_plugins/connector.py
@@ -71,7 +71,7 @@ class ConnectorPlugins(Connector):
                     "meaning": "JSONArrayMeaning",
                 },
                 {"name": "isDev", "type": "boolean", "meaning": "Boolean"},
-                {"name": "total_usages", "type": "integer", "meaning": "LongMeaning"},
+                {"name": "total_usages", "type": "bigint", "meaning": "LongMeaning"},
             ]
         }
